Extract Node-safe transaction display metadata
`CategoriesActions.ts` held five hundred lines deciding what a transaction
should be called: the category, the payee, the direction, and the label for
each action type. All of it is a pure function of the transaction, the wallet
and the account, but it sat behind Redux imports, so nothing outside the app
could ask the same question and get the same answer.

`src/util/txDisplay/` holds that logic now — `displayInfo` for the derivation,
`category` for the category strings, `txActionLabels` for the action names, and
`currencyCodes` for the ticker lookups. `CategoriesActions` re-exports what the
GUI already imported, so no scene changed.

The point is that two callers cannot drift. A transaction rendered in a list,
exported to CSV, or printed by a script now describes itself identically,
because it is the same code deciding.
